requirements.txt
streamlit==1.35.0
selenium==4.18.1
beautifulsoup4==4.12.3
requests==2.31.0
google-api-python-client==2.165.0
firebase-admin==6.7.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
google-generativeai==0.8.0
pandas==2.2.0
matplotlib==3.8.3
python-multipart==0.0.9 